requirements.txt
Flask==1.0.2
google-cloud-storage==1.7.0
gunicorn==19.8.1
WTForms-Appengine==0.1
numpy==1.15.2
keras==2.2.4
Pillow==5.3.0
tensorflow==1.12.0
